Move `retrySpawnProcess` to `async` module, to not have interfering mocking tests
